Fixed a bug in 'PostEffect'; Removed 'spatial_effect' from 'PostEffect'; Fixed a bug in 'Renderer'.
Fixed the bug that leads to incorrect color processing when fxaa is on in 'PostEffect'. It is caused by the buffer data swapping and incorrect usage of the buffer. Removed 'spatial_effect' for clearer logic. The user can certainly still apply this effect by adding a separate pass. This should not be too problematic performance-wise, as these effects themselves tend to be expensive so another sequence iteration should not be the ottleneck.

+++ termmy/rendering/renderer.py
@@ -120,7 +120,7 @@ def _initiate_unmatched_color_buffer(base: ColorBuffer,
     for y in range(out_height):
         for x in range(out_width):
             i = row_starting_index + x
-            color = base_data[int(y_scale*y*base_width + x_scale*x)]
+            color = base_data[int(y_scale*y)*base_width + int(x_scale*x)]
             old_color = out.data[i]
             old_color.r = color.r
             old_color.g = color.g
@@ -163,7 +163,7 @@ def _initiate_unmatched_ms_buffer(base: ColorBuffer,
     sample_num = len(msaa)
     for y in range(out_height):
         for x in range(out_width):
-            color = base_data[int(y_scale*y*base_width + x_scale*x)]
+            color = base_data[int(y_scale*y)*base_width + int(x_scale*x)]
             start = (row_starting_index + x) * sample_num
             for j in range(sample_num):
                 old_color = out.data[start + j]
